Update construct_settings overrides
Based on the discussion of #95 and googleapis/toolkit#70 --
this changes the interface of construct_settings a bit.

The new interface does not have the override parameters for
retrying and bundling anymore. Instead, it has a single
config_override that has the same strucure as client_config has.

API users can specify a copy of default config data with a few
modifications, but if the user-supplied config does not specify
some data, this will fall back to the default config, so
users can safely supply a part of the client config, just for
their edits.